node 0.10+ support
